feat(mcp): migrate host-agnostic config types + path-based load/save to dasclaw_mcp (F3.2 phase 2 PR 5a)
Move portable surface of MCP server config (types, validation, explicit-path
JSON load/save, is_localhost_url, EffectiveTransport, all portable tests)
from desktop-client/ironclaw/src/tools/mcp/config.rs (1214 lines) to the
new crates/dasclaw_mcp/src/config.rs (1084 lines).

The ironclaw file collapses to a 152-line shim that pub-re-exports the
portable surface and keeps only the host-specific wrappers: default
config path (rooted at dasclaw_base_dir()/mcp-servers.json), 5 no-path
wrappers, and 4 DB-backed wrappers that depend on crate::db::Database.

is_localhost_url is promoted from pub(crate) to pub so ironclaw's
auth.rs:366 can keep calling it across the crate boundary.

All call sites in extensions/manager.rs (5) and app.rs (1) are unchanged.

Order note: original plan was PR 5 = auth, PR 6 = config; swapped because
auth.rs has 15 usages of McpServerConfig and would otherwise create a
cross-crate cycle.

Tests: dasclaw_mcp 68/68 PASS (incl. 4 migrated tokio async fs tests),
       dasclaw cargo check OK.
Gates: fmt / clippy / check_no_panics / check_no_new_ironclaw_literal all green.
Review: 0 BLOCKER, 2 NOTE, 4 NIT (cleanup follow-up).

Closes #641 (F3.2 phase 2 PR 5a of N).
