+++ crates/dasclaw_mcp/Cargo.toml
@@ -10,6 +10,19 @@ publish = false
 path = "src/lib.rs"
 
 [dependencies]
+# `dasclaw_tool::ToolError` is the application-layer error returned by
+# config helpers (added in F3.2 phase 2 PR 5a, #641).
+dasclaw_tool = { path = "../dasclaw_tool" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = "1"
+# Async filesystem I/O for `load_mcp_servers_from` / `save_mcp_servers_to`.
+tokio = { version = "1", features = ["fs"] }
+# URL parsing for the localhost / loopback detection in `config::is_localhost_url`.
+url = "2"
+# RFC 9110 header name/value validation in `McpServerConfig::validate`.
+reqwest = { version = "0.12", default-features = false }
+
+[dev-dependencies]
+tempfile = "3"
+tokio = { version = "1", features = ["fs", "macros", "rt"] }
